refactor(View): rename `ui` directory to `view`
The `ui` directory contains the functions that build the screen which
belongs to the View component in the MVVM architecture.

Replace the old name with `views`.
