+++ src/macros.rs
@@ -26,7 +26,7 @@ macro_rules! loading_screen {
 
             let handle = std::thread::spawn(move || {
                 while loading_clone.load(std::sync::atomic::Ordering::Relaxed) {
-                    terminal = $crate::ui::loading_screen::render(terminal, $title);
+                    terminal = $crate::views::loading_screen::render(terminal, $title);
                     std::thread::sleep(std::time::Duration::from_millis(200));
                 }
