+++ src/views/details_actions.rs
