+++ src/views/latest.rs
