+++ src/views/navigation_bar.rs
