+++ src/views/popup/info_popup.rs
